pyproject.toml
[project]
name = "app-vitals-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dan McAulay", email = "dmcaulay@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.8.1",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]

[project.scripts]
mcp-server-toggl = "app_vitals_mcp.servers.toggl.main:main"
mcp-server-trello = "app_vitals_mcp.servers.trello.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-httpx>=0.35.0",
    "respx>=0.22.0",
    "ruff>=0.11.13",
]
